Update factory_girl to factory_bot (#25)
* Update factory_girl to factory_bot

The `factory_girl` gem is no longer available under that name, so
testing with it is now problematic. This change updates to the new name,
and updates other dependencies necessary for successful tests.

* Update static factory attributes to dynamic

factory_bot will no longer support static attributes in the next version
so this resolves a deprecation warning.